Improve pypowsybl converter structure and documentation
