+++ pandapower/test/converter/test_from_pypowsybl.py
@@ -1,3 +1,9 @@
+# -*- coding: utf-8 -*-
+
+# Copyright (c) 2016-2026 by University of Kassel and Fraunhofer Institute for Energy Economics
+# and Energy System Technology (IEE), Kassel. All rights reserved.
+
+
 import logging
 from pathlib import Path
 from typing import Any
